% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterCon.R
\name{InterCon}
\alias{InterCon}
\title{InterCon}
\usage{
InterCon(mat1, mat2, mat12, inter12, lambda0, rseq, nodes1, nodes2)
}
\arguments{
\item{mat1}{community Vc}

\item{mat2}{community Vc'}

\item{mat12}{community V(c, c')}

\item{lambda0}{A hyper parameter in objective function}

\item{rseq}{A sequence of cutoffs}

\item{nodes1}{nodes of the community c}

\item{nodes2}{nodes of the community c'}
}
\value{
result a list of 3 elements: rbest, interEdge and updated mat12
rbest is the best threshold.
InterEdge is a matrix including all significant edges
mat12 is the updated matrix after filtering by threshold.
}
\description{
InterCon
}
