% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLtest.R
\name{KL}
\alias{KL}
\title{KL Divergence Calculation}
\usage{
KL(y1, y2, width)
}
\arguments{
\item{y1}{distribution 1 (vector)}

\item{y2}{distribution 2 (vector)}

\item{width}{bin width. Default is 0.001}
}
\value{
KL distance between two distributions
}
\description{
\code{KL} Calculates the KL divergence between two distributions y1 and y2
}
\examples{

}
