% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLtest.R
\name{KLtest}
\alias{KLtest}
\title{KL test}
\usage{
KLtest(s, null, true_dist, a = 0.05, width = 0.001)
}
\arguments{
\item{s}{G(Vc, Vc') matrix distribution to test}

\item{null}{vector to sample null distributions (singleton indices)}

\item{true_dist}{Singleton matrix}

\item{a}{Significant level alpha. Default is 0.05}

\item{width}{default is 0.001. Can take out if you don't know what it is}
}
\value{
Returns \code{A} Boolean, 1 if KL-divergence significant (C and C' connected)
and 0 if divergence not significant.
}
\description{
\code{KLtest} Tests for the significance of KL divergence between
G(V_c, V_c') and G_R matrices
}
\examples{

}
