% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NICER}
\alias{NICER}
\title{NICE post processing}
\usage{
NICER(NICE)
}
\arguments{
\item{NICE}{A list object; result from NICE}
}
\value{
Returns a list of: K = number of clusters;
K.ind = indices of clusters
}
\description{
\code{NICER} process NICE object for step 2 analysis
}
